fix(client): 유저 이름에 정규식을 추가합니다 (#162)
